+++ scripts/run-backup.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
+
+# SPDX-FileCopyrightText: 2025-2026 Per Jensen
+#
 # SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of dar-backup-image:
+# https://github.com/per2jensen/dar-backup-image
+#
+# License terms and warranty disclaimer:
+# https://github.com/per2jensen/dar-backup-image/blob/main/LICENSE
 #
 # This script is part of the dar-backup-image project, which provides a Docker
 # image for running `dar` backups with sensible defaults and easy configuration.
@@ -92,7 +101,7 @@
 #
 #   #Check the command log:
 #   cat "$WORKDIR"/backups/dar-backup-commands.log
-#   
+#
 set -euo pipefail
 
 if ! command -v docker &>/dev/null; then