+++ scripts/save-dar-backup-image.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
+
+# SPDX-FileCopyrightText: 2025-2026 Per Jensen
+#
 # SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of dar-backup-image:
+# https://github.com/per2jensen/dar-backup-image
+#
+# License terms and warranty disclaimer:
+# https://github.com/per2jensen/dar-backup-image/blob/main/LICENSE
 #
 # save-dar-backup-image.sh — check if the latest dar-backup Docker image has been
 # archived locally, and if not, pull it from Docker Hub and save it as a