+++ src/clonepulse/util.py
@@ -1,4 +1,12 @@
+# SPDX-FileCopyrightText: 2025 Per Jensen
+#
 # SPDX-License-Identifier: MIT
+#
+# This file is part of dar-backup-image:
+# https://github.com/per2jensen/dar-backup-image
+#
+# License terms and warranty disclaimer:
+# https://github.com/per2jensen/dar-backup-image/blob/main/LICENSES/MIT.txt
 
 """
 Source code is here: https://github.com/per2jensen/clonepulse